+++ src/signalling/room.rs
@@ -574,12 +574,10 @@ impl Handler<CommandMessage> for Room {
                 sdp_answer,
             } => self.handle_make_sdp_answer(peer_id, sdp_answer),
             Command::SetIceCandidate { peer_id, candidate } => {
-                // TODO: add e2e test
+                // TODO: add E2E test
                 if candidate.candidate.is_empty() {
                     warn!("Empty candidate from Peer: {}, ignoring", peer_id);
-                    let fut: Box<
-                        dyn ActorFuture<Actor = _, Item = _, Error = _>,
-                    > = Box::new(actix::fut::ok(()));
+                    let fut: ActFuture<_, _> = Box::new(actix::fut::ok(()));
                     Ok(fut)
                 } else {
                     self.handle_set_ice_candidate(peer_id, candidate)